backend/requirements.txt
fastapi
uvicorn
kiteconnect
pandas
numpy
requests
python-dotenv